pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-framework-demo"
version = "0.1.0"
description = "A demo project for Semantic Kernel agent framework."
requires-python = ">=3.12"
dependencies = [
    "semantic-kernel",
    "python-dotenv",
    "azure-identity",
    "azure-ai-agents==1.1.0b4",
]

[project.scripts]
openapi = "openapi:run"
mcp = "mcp:run"

[tool.hatch.build.targets.wheel]
packages = ["."] 